Default to highp to fix issues in mobiles

+++ lib/src/shader/fragment.ts
@@ -28,7 +28,7 @@ export class TgdShaderFragment {
     public mainCode: TgdCodeBloc
 
     constructor({
-        precision = "mediump",
+        precision = "highp",
         uniforms = {},
         outputs = {
             FragColor: "vec4",
